requirements.txt
dj-database-url==0.5.0
Django==2.1
django-ckeditor==5.6.1
django-haystack==2.8.1
django-js-asset==1.1.0
django-machina==0.7.0
django-mptt==0.9.1
django-widget-tweaks==1.4.2
gunicorn==19.9.0
markdown2==2.3.5
Pillow==5.2.0
pytz==2018.5
virtualenv==16.0.0
whitenoise==4.1
Whoosh==2.7.4
